+++ xill-processor/src/main/java/nl/xillio/xill/plugins/date/services/DateServiceImpl.java
@@ -239,4 +239,14 @@ public class DateServiceImpl implements DateService, DateFactory {
             return pascalName;
         }
     }
+
+    @Override
+    public boolean isBefore(Date date1, Date date2) {
+        return date1.getZoned().isBefore(date2.getZoned());
+    }
+
+    @Override
+    public boolean isAfter(Date date1, Date date2) {
+        return date1.getZoned().isAfter(date2.getZoned());
+    }
 }