+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 * Make non-robot files editable in Xill IDE [CTC-1617]
 * File.move construct [CTC-1643]
 * Validate Xill robot when uploading to Xill Server [XSVR-21]
+* Add Date.isBefore and Date.isAfter constructs [CTC-1672]
 
 ### Change